Separated exceptions into different files
