+++ src/dotnet-uninstall/Messages.cs
@@ -42,7 +42,7 @@
         public static readonly string RequiredArgMissingForUninstallCommandExceptionMessage = "Required argument missing for the uninstall command";
         public static readonly string OptionsConflictExceptionMessageFormat = "Please specify only one of the options: {0}, {1}";
         public static readonly string CommandArgOptionConflictExceptionMessageFormat = "Please use either command arguments or the specified option: {0}";
-        public static readonly string InvalidInputVersionStringExceptionMessageFormat = "The specified version is not valid: {0}";
+        public static readonly string InvalidInputVersionExceptionMessageFormat = "The specified version is not valid: {0}";
         public static readonly string SpecifiedVersionNotFoundExceptionMessageFormat = "The specified version is not found: {0}";
     }
 }
