+++ src/dotnet-uninstall/Shared/Exceptions/BundleTypeMissingException.cs
@@ -1,0 +1,9 @@
+﻿namespace Microsoft.DotNet.Tools.Uninstall.Shared.Exceptions
+{
+    internal class BundleTypeMissingException : DotNetUninstallException
+    {
+        public BundleTypeMissingException() :
+            base(Messages.BundleTypeNotSpecifiedExceptionMessage)
+        { }
+    }
+}
