+++ src/dotnet-uninstall/Shared/Exceptions/CommandArgOptionConflictException.cs
@@ -1,0 +1,11 @@
+﻿using System.CommandLine;
+
+namespace Microsoft.DotNet.Tools.Uninstall.Shared.Exceptions
+{
+    internal class CommandArgOptionConflictException : DotNetUninstallException
+    {
+        public CommandArgOptionConflictException(Option option) :
+            base(string.Format(Messages.CommandArgOptionConflictExceptionMessageFormat, $"--{option.Name}"))
+        { }
+    }
+}
