+++ src/dotnet-uninstall/Shared/Exceptions/LinuxNotSupportedException.cs
@@ -1,0 +1,9 @@
+﻿namespace Microsoft.DotNet.Tools.Uninstall.Shared.Exceptions
+{
+    internal class LinuxNotSupportedException : DotNetUninstallException
+    {
+        public LinuxNotSupportedException() :
+            base(Messages.LinuxNotSupportedExceptionMessage)
+        { }
+    }
+}
