+++ src/dotnet-uninstall/Shared/Exceptions/RequiredArgMissingForUninstallCommandException.cs
@@ -1,0 +1,9 @@
+﻿namespace Microsoft.DotNet.Tools.Uninstall.Shared.Exceptions
+{
+    internal class RequiredArgMissingForUninstallCommandException : DotNetUninstallException
+    {
+        public RequiredArgMissingForUninstallCommandException() :
+            base(Messages.RequiredArgMissingForUninstallCommandExceptionMessage)
+        { }
+    }
+}
