+++ src/dotnet-uninstall/Shared/Exceptions/SpecifiedVersionNotFoundException.cs
@@ -1,0 +1,9 @@
+﻿namespace Microsoft.DotNet.Tools.Uninstall.Shared.Exceptions
+{
+    internal class SpecifiedVersionNotFoundException : DotNetUninstallException
+    {
+        public SpecifiedVersionNotFoundException(string versionString) :
+            base(string.Format(Messages.SpecifiedVersionNotFoundExceptionMessageFormat, versionString))
+        { }
+    }
+}
